Simple TCP listening, mostly for easy testing with nc. Set backend device properly

+++ backend.c
@@ -39,6 +39,8 @@
 #include <sys/queue.h>
 #include <fcntl.h>
 #include <stdarg.h>
+#include <netinet/in.h>
+#include <sys/stat.h>
 
 #include "status.h"
 #include "line.h"
@@ -67,6 +69,7 @@ struct backend_device {
 	char *client;
 	int client_fd;
 	struct event client_event;
+	int client_is_file;
 
 	struct status status;
 };
@@ -433,9 +436,9 @@ backend_error (struct bufferevent *be, short what, void *arg) {
 
 static void
 accept_connection (int fd, short what, void *cbarg) {
-	struct sockaddr_un unaddr;
-	socklen_t al = sizeof (unaddr);
-	int cfd = accept (fd, (struct sockaddr*)&unaddr, &al);
+	struct sockaddr_storage addr;
+	socklen_t al = sizeof (addr);
+	int cfd = accept (fd, (struct sockaddr*)&addr, &al);
 	struct backend *backend;
 
 	if (cfd < 0) {
@@ -449,6 +452,8 @@ accept_connection (int fd, short what, void *cbarg) {
 		close (cfd);
 		return;
 	}
+
+	backend->dev = cbarg;
 
 	backend->fd = cfd;
 	backend->be = bufferevent_new (cfd, backend_read, backend_write_done, backend_error, backend);
@@ -484,22 +489,52 @@ backend_listen_all (void) {
 	struct backend_device *bdev;
 
 	SLIST_FOREACH(bdev, &backends, link) {
-		struct sockaddr_un unaddr = {0};
+		struct sockaddr_storage addr = {0};
+		socklen_t al = 0;
+		char *e = NULL;
+		int p;
 
 		if (bdev->client_fd != -1 || !bdev->client)
 			continue;
 
-		if (unlink (bdev->client) && errno != ENOENT)
-			err(1, "unlink(%s)", bdev->client);
+		if ((p = strtol(bdev->client, &e, 0)) > 0 && p < 65536 && e && *e == '\0') {
+			struct sockaddr_in6 *in6addr = (struct sockaddr_in6*)&addr;
+			int one = 1;
 
-		bdev->client_fd = socket (PF_LOCAL, SOCK_STREAM, 0);
-		if (bdev->client_fd < 0)
-			err(1, "socket(%s)", bdev->client);
+			bdev->client_is_file = 0;
 
-		unaddr.sun_family = AF_LOCAL;
-		strlcpy (unaddr.sun_path, bdev->client, sizeof (unaddr.sun_path));
+			bdev->client_fd = socket (PF_INET6, SOCK_STREAM, 0);
+			if (bdev->client_fd < 0)
+				err(1, "socket(%s)", bdev->client);
 
-		if (bind (bdev->client_fd, (struct sockaddr*)&unaddr, sizeof (unaddr))) {
+			if (setsockopt(bdev->client_fd, SOL_SOCKET, SO_REUSEADDR, &one, sizeof(one)))
+				err(1, "setsockopt");
+
+			in6addr->sin6_family = AF_INET6;
+			in6addr->sin6_port = htons(p);
+			al = sizeof (*in6addr);
+		} else {
+			struct sockaddr_un *unaddr = (struct sockaddr_un*)&addr;
+			struct stat st;
+
+			bdev->client_is_file = 1;
+
+			if (!lstat(bdev->client, &st) && !S_ISSOCK(st.st_mode))
+				errx(1, "stat(%s): Exists but not socket");
+
+			if (unlink (bdev->client) && errno != ENOENT)
+				err(1, "unlink(%s)", bdev->client);
+
+			bdev->client_fd = socket (PF_LOCAL, SOCK_STREAM, 0);
+			if (bdev->client_fd < 0)
+				err(1, "socket(%s)", bdev->client);
+
+			unaddr->sun_family = AF_LOCAL;
+			strlcpy (unaddr->sun_path, bdev->client, sizeof (unaddr->sun_path));
+			al = sizeof (*unaddr);
+		}
+
+		if (bind (bdev->client_fd, (struct sockaddr*)&addr, al)) {
 			err(1, "bind(%s)", bdev->client);
 		}
 
@@ -524,7 +559,7 @@ backend_close_all (void) {
 
 		event_del (&bdev->client_event);
 		close (bdev->client_fd);
-		if (bdev->client)
+		if (bdev->client && bdev->client_is_file)
 			unlink (bdev->client);
 	}
 }
